backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
httpx>=0.25.0
openai-whisper>=20231117
pyttsx3>=2.90
openpyxl>=3.1.2
pillow>=10.0.0
pytesseract>=0.3.10
pypdf2>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
aiofiles>=23.2.1
pydantic>=2.4.0
python-dotenv>=1.0.0
